+++ test/Random.ts
@@ -1,14 +1,14 @@
 import * as assert from 'assert'
-import { random, randomInt, randomBool, randomRange } from '../src/Random'
+import * as _ from '../src/Random'
 
 describe('Random', () => {
   it('random', () => {
-    const n = random()
+    const n = _.random()
     assert.deepStrictEqual(typeof n, 'number')
   })
 
   it('randomInt', () => {
-    const n = randomInt(0, 10)()
+    const n = _.randomInt(0, 10)()
     assert.deepStrictEqual(typeof n, 'number')
     assert.deepStrictEqual(n % 1 === 0, true)
     assert.deepStrictEqual(n >= 0, true)
@@ -17,7 +17,7 @@ describe('Random', () => {
 
   it('randomRange', () => {
     for (let i = 0; i < 10; i++) {
-      const n = randomRange(0, 10)()
+      const n = _.randomRange(0, 10)()
       assert.deepStrictEqual(typeof n, 'number')
       assert.deepStrictEqual(n >= 0, true)
       assert.deepStrictEqual(n < 10, true)
@@ -25,7 +25,12 @@ describe('Random', () => {
   })
 
   it('randomBool', () => {
-    const b = randomBool()
+    const b = _.randomBool()
     assert.deepStrictEqual(typeof b, 'boolean')
+  })
+
+  it('randomElem', () => {
+    const e = _.randomElem([1, 2, 3])()
+    assert.deepStrictEqual(e >= 1 && e <= 3, true)
   })
 })